+++ app/src/main/java/eu/kanade/domain/source/service/SourcePreferences.kt
@@ -46,16 +46,6 @@ class SourcePreferences(
         emptySet(),
     )
 
-    fun invalidAnimeExtensions() = preferenceStore.getStringSet(
-        Preference.appStateKey("invalid_anime_extensions"),
-        emptySet(),
-    )
-
-    fun invalidMangaExtensions() = preferenceStore.getStringSet(
-        Preference.appStateKey("invalid_manga_extensions"),
-        emptySet(),
-    )
-
     fun globalSearchFilterState() = preferenceStore.getBoolean(
         Preference.appStateKey("has_filters_toggle_state"),
         false,