+++ app/src/main/java/eu/kanade/tachiyomi/extension/manga/util/MangaExtensionInstallReceiver.kt
@@ -55,7 +55,6 @@ internal class MangaExtensionInstallReceiver(private val listener: Listener) : B
                     when (val result = getExtensionFromIntent(context, intent)) {
                         is MangaLoadResult.Success -> listener.onExtensionInstalled(result.extension)
                         is MangaLoadResult.Untrusted -> listener.onExtensionUntrusted(result.extension)
-                        is MangaLoadResult.Invalid -> listener.onExtensionInvalid(result)
                         else -> {}
                     }
                 }
@@ -65,7 +64,6 @@ internal class MangaExtensionInstallReceiver(private val listener: Listener) : B
                     when (val result = loadWithRetryOnReplace(context, intent)) {
                         is MangaLoadResult.Success -> listener.onExtensionUpdated(result.extension)
                         is MangaLoadResult.Untrusted -> listener.onExtensionUntrusted(result.extension)
-                        is MangaLoadResult.Invalid -> listener.onExtensionInvalid(result)
                         else -> {}
                     }
                 }
@@ -138,7 +136,6 @@ internal class MangaExtensionInstallReceiver(private val listener: Listener) : B
         fun onExtensionInstalled(extension: MangaExtension.Installed)
         fun onExtensionUpdated(extension: MangaExtension.Installed)
         fun onExtensionUntrusted(extension: MangaExtension.Untrusted)
-        fun onExtensionInvalid(extension: MangaLoadResult.Invalid)
         fun onPackageUninstalled(pkgName: String)
     }